backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
edge-tts>=6.1.9
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.0.0
httpx>=0.26.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
